requirements.txt
# requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
piexif>=1.1.3